# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareOps AI - Inspection Readiness Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /training:
    get:
      operationId: listTraining
      tags: [training]
      summary: List all training records
      parameters:
        - name: staffId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [current, expiring_soon, expired]
      responses:
        "200":
          description: Training records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingRecord"
    post:
      operationId: createTrainingRecord
      tags: [training]
      summary: Record a training completion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingInput"
      responses:
        "201":
          description: Created training record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRecord"
  /training/{id}:
    patch:
      operationId: updateTrainingRecord
      tags: [training]
      summary: Update a training record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingUpdate"
      responses:
        "200":
          description: Updated training record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRecord"
...
components:
  schemas:
...
    TrainingRecord:
      type: object
      required: [id, staffId, staffName, trainingType, category, completedDate, expiryDate, status]
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: string
        trainingType:
          type: string
        category:
          type: string
          enum: [mandatory, specialist, refresher]
        completedDate:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum: [current, expiring_soon, expired]
        notes:
          type: ["string", "null"]
    TrainingInput:
      type: object
      required: [staffId, trainingType, category, completedDate, expiryDate]
      properties:
        staffId:
          type: integer
        trainingType:
          type: string
        category:
          type: string
          enum: [mandatory, specialist, refresher]
        completedDate:
          type: string
        expiryDate:
          type: string
        notes:
          type: string
    TrainingUpdate:
      type: object
      properties:
        completedDate:
          type: string
        expiryDate:
          type: string
        notes:
          type: string
...